streaming: count highlights for diff/commit search
Closes #18301

Other options considered:

It is an overhead to run `ToCommitSearchResult` for every result. To
avoid this one could wrap each `CommitSearchResultResolver` in a struct
that reimplements `ResultCount`. Depending on whether streaming is
active or not we wrap or we don't. The best place to do this would be in
`doResults`. I compared the two implementations and found the one in
this PR cleaner and more localized to streaming.